Implement setA
Implements setAttribute

This makes it so the patcher/decoder now both support setAttribute
calls. removeAttribute is not yet supported, but wanted to get this in
first.
