+++ doc/user-profile.html
@@ -53,7 +53,8 @@
 	}).observe(sourceDoc.documentElement, {
 		characterData: true,
 		childList: true,
-		subtree: true
+		subtree: true,
+		attributes: true
 	});
 
 </script>
